moved fuzzy search into utils to be accessible to other components
